+++ src/components/narrative/renderNarrativeToggle.js
@@ -17,7 +17,7 @@ export const renderNarrativeToggle = (dispatch, narrativeIsDisplayed) => {
         cursor: "pointer",
         color: darkGrey
       }}
-      onClick={() => dispatch({type: TOGGLE_NARRATIVE, display: !narrativeIsDisplayed})}
+      onClick={() => dispatch({type: TOGGLE_NARRATIVE, narrativeOn: !narrativeIsDisplayed})}
     >
       {narrativeIsDisplayed ? "explore the data yourself" : "return to the narrative"}
     </button>